refactor: change folder structure
Breaking out files into multiple logical elements
