+++ app/items.py
@@ -1,0 +1,11 @@
+from scrapy import Item, Field
+
+
+class WebsiteItem(Item):
+
+    """
+    Defines the item output
+    """
+
+    company_url = Field()
+    company_text = Field()
